matrix.c: Fix a cross section reading bug.
When the value in the first column is outside the analysis window's
selected range (xmin, xmax), we should skip forward to the end of the
line.  Otherwise we treat cross section values as wavelengths, which
causes a problem when the cross section value happens to be in the
same range as the wavelengths.