coda vcpkg port: Remove xdr.dll from HDF4_DLLS, cleanup.
xdr.dll is NOTFOUND for HDF4 4.3.0 and this causes installation to
fail on windows.

We can pass the location of Flex and Bison in an environment variable
if we add that variable to VCPKG_KEEP_ENV_VARS.

We don't use coda's HDF5 functions, so remove that dependency from our
portfile.