matrix_netcdf_read.cpp: Avoid segfault from bad wavelength range.
When trying to read data with a range (xmin, xmax) which lies
completely outside of the wavelength domain of the provided matrix
file, the selected number of wavelengths becomes == 1, which trips up
the derivative calculation later on.  We therefore exclude this case
with an exception.

Also avoid some signed / unsigned conversion warning and remove unused
variables.